% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichR-package.r
\docType{data}
\name{background}
\alias{background}
\title{Example background genes}
\format{
vector
}
\usage{
data(background)
}
\description{
This is a character vector which consists of 20,625 gene symbols taken from the Enrichr website.
}
\examples{
data(background)
length(background)
}
\keyword{datasets}
