% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{barplot: a function for GO,KEGG enrich results visual in barplot}
\usage{
barplot(dat1,term_num = 6,fill_level=c("MF","BP","CC","KEGG","TF"),term_width=55)
barplot(dat1)
barplot(dat1,term_num = 8,fill_level=c("GO","KEGG","TF"),term_width=55)
barplot(dat1,term_num = 6,fill_level=c("MF","BP","CC"))
}
\arguments{
\item{fill_level}{vector. for show the terms' type.Default:c("MF","BP","CC","KEGG")
if you also enrich TF or other.you can use c("MF","BP","CC","KEGG","TF").
this vector's string must be the same with the input dataframe type row.
\code{fill_level} must be a vector,which is contained the input dataframe row \code{type}.}

\item{term_num}{numeric. for show terms count number. Default:8}

\item{term_width}{numeric. for show terms' description width. Default:40}

\item{data_GO_all.}{must be have 4 rows. and rownames must be \verb{ID Description count type}.
and the row of \code{count} must be number.
a dataframe like this
head(data_GO_all)
ID                    Description count type
GO:0015979                 photosynthesis    31   BP
GO:0009535 chloroplast thylakoid membrane    26   BP
GO:0055035     plastid thylakoid membrane    26   CC
GO:0042651             thylakoid membrane    28   MF}
}
\value{
a ggplot2 results
}
\description{
a function for GO,KEGG or TF enrich results visual in barplot
}
\examples{
load("data/enrich_result.Rdata")
barplot(dat1, term_num = 4)
barplot(dat1)
barplot(dat1, term_num = 5, term_width = 45)
barplot(dat1, term_num = 10, term_width = 50)
}
