% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{DE}
\alias{DE}
\title{Disease Ontology Enrichment analysis function}
\usage{
DE(df, keytype = "SYMBOL", padj.method = "BH", filename = NULL)
}
\arguments{
\item{df}{DGE files (DESeq2 result files) or vector contains gene names}

\item{keytype}{Keytype}

\item{filename}{output filename}
}
\description{
Disease Ontology Enrichment analysis function
}
\author{
Kai Guo
}
