% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.add_background}
\alias{.add_background}
\title{Upload background list using Speedrichr API}
\usage{
.add_background(genes)
}
\arguments{
\item{genes}{gene list}
}
\value{
R object from JSON
}
\description{
Upload background list using Speedrichr API
}
\details{
Upload background list using Speedrichr API
}
\author{
I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
