% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.add_list}
\alias{.add_list}
\title{Upload gene list using Speedrichr API}
\usage{
.add_list(genes)
}
\arguments{
\item{genes}{Input genes}
}
\value{
R object that corresponds to the JSON object
}
\description{
Upload gene list using Speedrichr API
}
\details{
Upload gene list using Speedrichr API
}
\author{
I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
