% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.enrichment_prep_df}
\alias{.enrichment_prep_df}
\title{Given a Enrichr output, order and subset criteria, returns a data frame accordingly}
\usage{
.enrichment_prep_df(df, showTerms = 20, orderBy = "P.value")
}
\arguments{
\item{df}{Enrichr output}

\item{showTerms}{Number of terms to show. Default 20.}

\item{orderBy}{Column for ordering. Default "P.value"}
}
\value{
Data frame
}
\description{
Given a Enrichr output, order and subset criteria, returns a data frame accordingly
}
\details{
Given a Enrichr output, order and subset criteria, returns a data frame accordingly
}
\author{
I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
