% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.formatGenes}
\alias{.formatGenes}
\title{FormatGenes}
\usage{
.formatGenes(x, type = c("standard", "background"))
}
\arguments{
\item{x}{Vector or dataframe of genes with or without score}

\item{type}{Depends on type of gene input}
}
\value{
Character vector
}
\description{
Given an input, check format and return a character vector
}
\details{
In standard analysis without background, crisp (symbols only) and
fuzzy (with scores) gene sets are acceptable
In analysis with background, only crisp gene sets are acceptable
}
\author{
I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
