% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.get_backgroundenrich}
\alias{.get_backgroundenrich}
\title{Get enrichment result using Speedrichr API}
\usage{
.get_backgroundenrich(uId, bId, db)
}
\arguments{
\item{uId}{user List ID}

\item{bId}{background ID}

\item{db}{background Type}
}
\value{
R object from JSON
}
\description{
Get enrichment result using Speedrichr API
}
\details{
Get enrichment result using Speedrichr API
}
\author{
I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
