% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.onLoad}
\alias{.onLoad}
\title{onLoad hook to setup package options}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\value{
Nil
}
\description{
onLoad hook to setup package options
}
\details{
onLoad hook to setup package options and to check connection to website
}
\author{
Wajid Jawaid
}
