% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.proxyOpts}
\alias{.proxyOpts}
\title{Internal function to check RCurlOptions}
\usage{
.proxyOpts()
}
\value{
Named vector
}
\description{
Internal function to check RCurlOptions
}
\details{
Internal function to check RCurlOptions
}
\author{
I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
