% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.read_gmt}
\alias{.read_gmt}
\title{Download and parse GMT files from Enrichr}
\usage{
.read_gmt(db)
}
\arguments{
\item{db}{library}
}
\value{
List object
}
\description{
Download and parse GMT files from Enrichr
}
\details{
Download and parse GMT files from Enrichr
}
\author{
I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
