% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{dotplot: Combine multiple groups of GO/KEGG/TF analysis to generate a bubble chart}
\usage{
dotplot(data_GO_all, xlabels = data_GO_all$type, output, ranges = c(4, 7))
}
\arguments{
\item{data_GO_all}{a dataframe,contain columns name of "type","Description","adjust","Count"}

\item{xlabels}{a vector for x-axis labels}

\item{output}{output file prefix}

\item{ranges}{Default: c(4,7); a range,like c(3,5), for control the size of the bubble}
}
\value{
a ggplot object
}
\description{
dotplot: Combine multiple groups of GO/KEGG/TF analysis to generate a bubble chart
}
\examples{
load("data/data_GO_all.Rdata")
dotplot(data_GO_all,xlabels=xlabels,output="All_GO")
dotplot(data_GO_all,xlabels=xlabels,output="All_GO",range=c(3,5))
dotplot(data_GO_all,xlabels=xlabels,output="All_GO",range=c(4,7))
}
