% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich}
\alias{enrich}
\title{Enrichment analysis for any type of annotation data}
\usage{
enrich(df, annot, annot.info = NULL, minSize = 1, maxSize = 500,
  keepRich = TRUE, filename = NULL, gene.cutoff = 0.01,
  padj.method = "BH")
}
\arguments{
\item{df}{DGE files (DESeq2 result files) or vector contains gene names}

\item{annot}{annotation file for all genes}

\item{annot.info}{Term of all annotation}

\item{keepRich}{keep terms with high rich factor}

\item{filename}{output filename}

\item{gene.cutoff}{DGE singificant cutoff value}

\item{padj.method}{p value adjust method}
}
\description{
Enrichment analysis for any type of annotation data
}
\author{
Kai Guo
}
