% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichbar}
\alias{enrichbar}
\title{Display enrichment result By using barchart}
\usage{
enrichbar(resultFis, top = 50, pvalue.cutoff = 0.05,
  padj.cutoff = NULL, order = FALSE, fontsize.x = 10,
  fontsize.y = 10, fontsize.text = 3, angle = 75, usePadj = TRUE,
  filename = NULL)
}
\arguments{
\item{resultFis:}{Ennrichment analysis result data.frame}

\item{top:}{Number of Terms you want to display}

\item{pvalue.cutoff:}{the cut-off value for selecting Term}

\item{padj.cutoff:}{the padj cut-off value for selecting Term}
}
\description{
Display enrichment result By using barchart
}
\author{
Kai Guo
}
