% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichR-package.r
\docType{package}
\name{enrichR-package}
\alias{enrichR}
\alias{enrichR-package}
\title{enrichR package}
\description{
The enrichR package provides an R interface to all enrichR (https://maayanlab.cloud/Enrichr/) databases.
}
\author{
Wajid Jawaid \email{wajid.jawaid@gmail.com}
}
\keyword{internal}
