% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE.R
\name{GE.plot}
\alias{GE.plot}
\title{Display GO enrichment result}
\usage{
GE.plot(resultFis, top = 50, pvalue.cutoff = 0.05, order = FALSE,
  fontsize.x = 10, fontsize.y = 10, fontsize.text = 3, angle = 75,
  padj.cutoff = NULL, usePadj = TRUE, filename = NULL)
}
\arguments{
\item{resultFis:}{GO ennrichment analysis result data.frame}

\item{top:}{Number of Terms you want to display}

\item{filenam:}{output filename}

\item{pvalue.cutoff:}{the cut-off value for selecting Term}

\item{padj.cutoff:}{the padj cut-off value for selecting Term}
}
\description{
Display GO enrichment result
}
\author{
Kai Guo
}
