% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{gene2venn}
\alias{gene2venn}
\title{gene2venn: input gene ,out put venn and upset image.}
\usage{
gene2venn(data_gene, outname = "demo")
}
\arguments{
\item{data_gene}{dataframe. contain the columns of each group.}

\item{outname}{string.Default:"demo",for output file name prefix.}
}
\value{
the common venn gene
}
\description{
gene2venn: input gene ,out put venn and upset image.
}
\note{
The output file is \code{outname_matrix.csv},\code{outname_venn.pdf},\code{outname_upset.pdf}.
}
\examples{
load("venn.Rdata")
head(data_venn)
  PH6WC_XY335   PH4CVvsXY335   PH4CVvsPH6WC         RIL335       F2_3_335       F2_3_958         RIL958
1 Zm00001d033896 Zm00001d044327 Zm00001d033896 Zm00001d053675 Zm00001d031899 Zm00001d034543 Zm00001d018779
2 Zm00001d032359 Zm00001d012313 Zm00001d004348 Zm00001d043045 Zm00001d013795 Zm00001d045451 Zm00001d002000
3 Zm00001d008977 Zm00001d035030 Zm00001d009840 Zm00001d047789 Zm00001d021310 Zm00001d021784 Zm00001d033091
4 Zm00001d038546 Zm00001d013456 Zm00001d012313 Zm00001d018386 Zm00001d047789 Zm00001d047789 Zm00001d043299
gene2venn(data_venn)

load("../data/veen.Rdata")
intersect_gene <- gene2venn(data_venn \%>\% select(1:4)) #鑾峰彇鎵€鏈夊垪鐨勪氦闆㠼㸶
write.table(intersect_gene,file = "intersect_gene.csv",row.names = FALSE)
}
