% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichR-package.r
\docType{data}
\name{genes790}
\alias{genes790}
\title{790 gene symbols}
\format{A vector.}
\usage{
genes790
}
\description{
This is a character vector which consists of randomly selected 790 genes.
}
\keyword{datasets}
