% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{get_matrix}
\alias{get_matrix}
\title{Title get_matrix:return a 0/1 matrix dataframe.}
\usage{
get_matrix(data_venn, getlogic = FALSE)
}
\arguments{
\item{data_venn}{dataframe.input the genelist dataframe}

\item{getlogic}{TRUE or FALSE.Default:FALSE.if set true, the output will be logic dataframe.}
}
\value{
dataframe. Default:contain 0 or 1 dataframe.
}
\description{
Title get_matrix:return a 0/1 matrix dataframe.
}
\examples{
load("venn.Rdata")
head(data_venn)
  PH6WC_XY335   PH4CVvsXY335   PH4CVvsPH6WC         RIL335       F2_3_335       F2_3_958         RIL958
1 Zm00001d033896 Zm00001d044327 Zm00001d033896 Zm00001d053675 Zm00001d031899 Zm00001d034543 Zm00001d018779
2 Zm00001d032359 Zm00001d012313 Zm00001d004348 Zm00001d043045 Zm00001d013795 Zm00001d045451 Zm00001d002000
3 Zm00001d008977 Zm00001d035030 Zm00001d009840 Zm00001d047789 Zm00001d021310 Zm00001d021784 Zm00001d033091
4 Zm00001d038546 Zm00001d013456 Zm00001d012313 Zm00001d018386 Zm00001d047789 Zm00001d047789 Zm00001d043299
get_matrix(data_venn)
}
