% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{get_venn}
\alias{get_venn}
\title{get_venn: input a dataframe ,output venn diagram.Only draw venn graphs of 2-4 dimensions}
\usage{
get_venn(data2, fill_color = color, percentage = TRUE)
}
\arguments{
\item{fill_color}{Default:color , can be color 1 to color4, or a color vector of length 4}

\item{percentage}{logic:TRUE or FALSE, control show the percentage .}

\item{data2:a}{o/1 data matrix. output from \code{get_matrix}.}
}
\value{
return a ggplot image.
}
\description{
get_venn: input a dataframe ,output venn diagram.Only draw venn graphs of 2-4 dimensions
}
\examples{
load("venn.Rdata")
data2 <- get_matrix(data_gene)
get_venn(data2,fill_color=color1,percentage=FALSE)
get_venn(data2,fill_color=color4)
get_venn(data2,fill_color=color3,percentage=FALSE)
get_venn(data2,fill_color=color1,percentage=TRUE)
}
