% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdetail.R
\name{getdetail}
\alias{getdetail}
\title{Get detail information from enrichment results}
\usage{
getdetail(rese, resd)
}
\arguments{
\item{rese}{Functional enrichment results}

\item{resd}{DEG result with one gene name column or vector of input genes}
}
\description{
Get detail information from enrichment results
}
\author{
Kai Guo
}
