% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEnrichr}
\alias{getEnrichr}
\title{Helper function for GET}
\usage{
getEnrichr(url, startUp = FALSE, ...)
}
\arguments{
\item{url}{(Required). URL address requested}

\item{startUp}{(Optional). Check if function is called during startUp}

\item{...}{(Optional). Additional parameters to pass to GET}
}
\value{
same as GET
}
\description{
Helper function
}
\details{
Helper function for GET
}
\author{
Wajid Jawaid \email{wajid.jawaid@gmail.com}

I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
