% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getPCA}
\alias{getPCA}
\title{get_pca}
\usage{
getPCA(dat, group_list)
}
\arguments{
\item{dat}{a dataframe; row is samples/materials,column is genename}

\item{group_list}{a list: contain the samples/materials repeat.
group_list should like this:\verb{"RC"  "RC"  "RC"  "RT1" "RT1" "RT1" "RT2" "RT2" "RT2" "RT3" "RT3" "RT3"}}
}
\value{
a ggplot2 image.
}
\description{
get_pca
}
\examples{
load("PCA.Rdata")
group_level <- substr(colnames(read_counts), 1, 3)
getPCA(read_counts,group_level)
}
