% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnet.R
\name{gnet}
\alias{gnet}
\title{Draw network with ggplot2 style}
\usage{
gnet(df, rhs, top = 50, pvalue.cutoff = 0.05, padj.cutoff = NULL,
  weightcut = 0.2, useTerm = TRUE, writeCyt = FALSE,
  cytoscapeFile = "network-file-for-cytoscape.txt",
  vertex.label.color = "black", vertex.label.cex = 0.5,
  vertex.node.shape = NULL, layout = layout.fruchterman.reingold,
  savefig = FALSE, filename = "network", width = 7, height = 7,
  segment.size = 0.2, ...)
}
\arguments{
\item{df}{Differential expression result or vector of genes}

\item{rhs}{Enrichment results}

\item{top}{number of terms to show (default: 50)}

\item{pvalue.cutoff}{cutoff p value for enrichment result}

\item{padj.cutoff}{cutoff p adjust value for enrichment result}

\item{weightcut}{cutoff value for the edges}

\item{useTerm}{use terms for nodes (default: TRUE)}

\item{writeCyt}{write out the cytoscape file}

\item{cytoscapeFile}{output cytoscape File}

\item{vertex.label.color}{label color}

\item{vertex.label.cex}{label size}

\item{vertex.node.shape}{vector of shape(1..24) and names of the vector should be the terms}

\item{layout}{layout method}

\item{savefig}{save figures or not}

\item{filename}{output figure name}

\item{width}{width for output figure}

\item{height}{height for output figure}

\item{segment.size}{size for label segment}
}
\description{
gnet provides method to draw the enrichment results with a ggplot2 style network
}
\author{
Kai Guo
}
