% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GOenrich}
\alias{GOenrich}
\title{GOenrich: for GO enrich analysis
GOenrich ,KEGG_enrich,TF_enrich, have the some type input and output}
\usage{
GOenrich(genelist, genetype, GOfile = "GO.info")
}
\arguments{
\item{genelist}{a gene list like this"
MELO3C000003
MELO3C000548
MELO3C001803}

\item{genetype}{a string like "WTvsT1" or "Mut12vsWT" for the output image file prefix}

\item{GOfile}{a filename "GO.info",default:"Go.info".GO.info must be only have 4 column,not have header. separator must is tab.
"GO.info" Demo:
#GO:0016020    membrane    MELO3C000003    cellular_component
#GO:0016021    integral    MELO3C000003    cellular_component}
}
\value{
return GO enrich results
}
\description{
GOenrich: for GO enrich analysis
GOenrich ,KEGG_enrich,TF_enrich, have the some type input and output
}
\examples{
genelist <- read.delim("data/ST1.txt")
GOenrich(genelist,"WTvsT1","data/GO.info")
GOenrich(genelist,"WTvsT1")
go_res <- GOenrich("gene.xls","WTvsT1")
if gO_res$Count >1,mean GO enrich is successful.
if go_res$Count >1,mean no gene can be enrich.

}
