% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{gsea}
\alias{gsea}
\title{Enrichment analysis for any type of annotation data}
\usage{
gsea(x, annot, annot.info = NULL, minSize = 15, maxSize = 500,
  nperm = 5000, filename = NULL, padj.method = "BH", table = FALSE)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{annot}{annotation file for all genes}

\item{annot.info}{Term with annotation details}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}

\item{filename}{output filename}

\item{padj.method}{p value adjust method}
}
\description{
Enrichment analysis for any type of annotation data
}
\author{
Kai Guo
}
