% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org.R
\name{idconvert}
\alias{idconvert}
\title{Convert ID between ENTREZID to SYMBOL or other type ID based on bioconductor annotation package}
\usage{
idconvert(species, keys, fkeytype, tkeytype)
}
\arguments{
\item{species}{you can check the support species by using showData()}

\item{keys}{input genes}

\item{fkeytype}{the gene type you want to convert}

\item{tkeytype}{the gene type you want to get}
}
\description{
Convert ID between ENTREZID to SYMBOL or other type ID based on bioconductor annotation package
}
\author{
Kai Guo
}
