% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE.R
\name{IE.ensemble}
\alias{IE.ensemble}
\title{InterPro Enrichment analysis function}
\usage{
IE.ensemble(df, IO_FILE, gene.cutoff = 0.01, minSize = 2,
  maxSize = 500, keepRich = TRUE, padj.method = "BH",
  filename = NULL, cutoff = 0.05)
}
\arguments{
\item{df}{DGE files (DESeq2 result files) or vector contains gene names}

\item{IO_FILE}{InterPro annotation data}

\item{gene.cutoff}{the cut-off value for select DEGs (default: 0.01)}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{padj.method}{p value adjust method (default: BH)}

\item{filename}{output filename}

\item{cutoff}{cutoff value for filtering significant terms (default: 0.05)}
}
\description{
InterPro Enrichment analysis function
}
\author{
Kai Guo
}
