% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichR-package.r
\docType{data}
\name{input}
\alias{input}
\title{Example input genes}
\format{
vector
}
\usage{
data(input)
}
\description{
This is a character vector which consists of 375 gene symbols taken from the Enrichr website.
}
\examples{
data(input)
length(input)
}
\keyword{datasets}
