% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KE.R
\name{KE.plot}
\alias{KE.plot}
\title{Display KEGG enrichment result}
\usage{
KE.plot(resultFis, pvalue.cutoff = 0.05, top = 50, order = FALSE,
  font.x = "bold", font.y = "bold", fontsize.x = 10,
  fontsize.y = 10, padj.cutoff = NULL, usePadj = TRUE,
  filename = NULL)
}
\arguments{
\item{resultFis:}{KEGG ennrichment analysis result data.frame}

\item{top:}{Number of Terms you want to display}

\item{filenam:}{output filename}

\item{pvalue.cutoff:}{the cut-off value for selecting Term}

\item{padj.cutoff:}{the padj cut-off value for selecting Term}
}
\description{
Display KEGG enrichment result
}
\author{
Kai Guo
}
