% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{KEGGenrich}
\alias{KEGGenrich}
\title{KEGGenrich: for enrich kegg}
\usage{
KEGGenrich(genelist, genetype, KEGGfile = "KEGG.info")
}
\arguments{
\item{genelist}{a list of gene}

\item{genetype}{a string}

\item{KEGGfile}{Default:KEGG.info}
}
\value{
KEGG the enrich results
}
\description{
KEGGenrich: for enrich kegg
}
\examples{
genelist <- read.delim("data/ST1.txt")
KEGGenrich(genelist,"ST1","data/KEGG.info")

}
