% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{makeesanno}
\alias{makeesanno}
\title{make annotation data except plant based on ENSEMBLE}
\usage{
makeesanno(species = "Human", host = "uswest.ensembl.org",
  ann_type = "GO")
}
\arguments{
\item{species}{the species you want to search, you can use showplant to get the species name}

\item{host}{the ensemble API host,for plant you can use plants.ensembl.org and for human and other species you can use uswest.ensembl.org}

\item{ann_type}{the type of function annotation(GO,KEGG,PFAM,InterPro) you want get from ensemble}
}
\description{
make annotation data except plant based on ENSEMBLE
}
\author{
Kai Guo
}
