% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org.R
\name{makeKOdat}
\alias{makeKOdat}
\title{make KEGG annotation data function}
\usage{
makeKOdat(species = "human", keytype = "ENTREZID", builtin = TRUE)
}
\arguments{
\item{species:}{you can check the support species by using showData()}

\item{keytype:}{the gene type}
}
\description{
make KEGG annotation data function
}
\author{
Kai Guo
}
