% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeown.R
\name{makeOwnGO}
\alias{makeOwnGO}
\title{Make user define annotation data}
\usage{
makeOwnGO(anno, godata = godata, use = TRUE)
}
\arguments{
\item{anno:}{data.frame includes gene_id and GO or Pathway information}

\item{godata:}{godata inlude BP,MF,CC annotation}

\item{use:}{check if the user choose to load internal godata}
}
\description{
Make user define annotation data
}
\author{
Kai Guo
}
