% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{makeplantann}
\alias{makeplantann}
\title{make plant annotation data based on ENSEMBLE}
\usage{
makeplantann(species = "Arabidopsis t", host = "plants.ensembl.org",
  ann_type = "GO")
}
\arguments{
\item{species}{the sepcies you want to search, you can use showplant to get the species name}

\item{host}{the ensemble API host,for plant you can use plants.ensembl.org and for human and other species you can use uswest.ensembl.org}

\item{ann_type}{the type of function annotation(GO,KEGG,PFAM,InterPro) you want get from ensemble}
}
\description{
make plant annotation data based on ENSEMBLE
}
\author{
Kai Guo
}
