% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrodata.R
\name{makeplantROdat}
\alias{makeplantROdat}
\title{make Plant Reactome annotation data function}
\usage{
makeplantROdat(species = "Arabidopsis_thaliana")
}
\arguments{
\item{species:}{you can check the supported species by using showAvailableRO and showAvailablePlants}
}
\description{
make Plant Reactome annotation data function
}
\author{
Kai Guo
}
