% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrodata.R
\name{makeROdata}
\alias{makeROdata}
\title{make Reactome annotation data function}
\usage{
makeROdata(species = "Homo_sapiens")
}
\arguments{
\item{species}{you can check the supported species by using showAvailableRO and showAvailablePlants}
}
\description{
make Reactome annotation data function
}
\author{
Kai Guo
}
