% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbr.R
\name{msigdbr}
\alias{msigdbr}
\alias{makeMSIGDB}
\title{Download database from Msigdb and prepare for enrichment analysis}
\usage{
makeMSIGDB(species = "human", keytype = "SYMBOL", subcategory = NULL,
  category = NULL, path = "./", save = FALSE)
}
\arguments{
\item{species}{the species for query}

\item{keytype}{the gene ID type}

\item{subcategory}{Gene Set subcategory}

\item{category}{Gene set category}

\item{path}{path to save the dataset}

\item{save}{save the dataset or not}
}
\description{
Download database from Msigdb and prepare for enrichment analysis
}
\author{
Kai Guo
}
