% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrapper.R
\name{mungeResponseContent}
\alias{mungeResponseContent}
\title{Munge the Enrichr API response so it'll squeeze neatly (if untidily) into a dataframe.}
\usage{
mungeResponseContent(response.content)
}
\arguments{
\item{response.content}{result of calling httr::content on the GET request to the Enrichr API, after submitting a list for enrichment.}
}
\description{
The response from the Enrichr API is a list of lists, where each nested list item represents an enriched
category. The 6th item of each category (i.e. response.content[[category.idx]][[6]]) corresponds to the
genes that overlapped with the gene set behind that category. This function bascically collapses that list of
genes into a single string.
}
\details{
I'm sorry you ever had to look at this.
}

