% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE.R
\name{PE.plant}
\alias{PE.plant}
\title{PFAM Enrichment analysis function based on ensemble plant annotation data}
\usage{
PE.plant(df, PO_FILE, gene.cutoff = 0.01, padj.method = "BH", minSize = 2,
  maxSize = 500, keepRich = TRUE, filename = NULL)
}
\arguments{
\item{df:}{DGE files (DESeq2 result files) or vector contains gene names}

\item{PO_FILE:}{PFAM annotation data, you can get the data by using makeplantann function}

\item{filename:}{output filename}

\item{gene.cutoff:}{DGE singificant cutoff value}
}
\description{
PFAM Enrichment analysis function based on ensemble plant annotation data
}
\author{
Kai Guo
}
