% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{plot_gsea}
\alias{plot_gsea}
\alias{plot_fgsea}
\title{plot the gsea result}
\usage{
plot_fgsea(term, x, annot)
}
\arguments{
\item{term}{the significant term}

\item{x}{a vector include all log2FC with gene name}

\item{annot}{annotation file for all genes}
}
\description{
plot the gsea result
}
\author{
Kai Guo
}
