% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{plotgsea}
\alias{plotgsea}
\title{plot the gsea result}
\usage{
plotgsea(x, term, annot, gseaRes = NULL, annot.info = NULL)
}
\arguments{
\item{x}{a vector include all log2FC with gene name}

\item{term}{the significant term}

\item{annot}{annotation file for all genes}

\item{annot.info}{Term with annotation details}
}
\description{
plot the gsea result
}
\author{
Kai Guo
}
