% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RE.R
\name{RE.plant}
\alias{RE.plant}
\title{Reactome Pathway Enrichment analysis function for plant}
\usage{
RE.plant(df, RO_FILE, gene.cutoff = 0.01, padj.method = "BH",
  minSize = 2, maxSize = 500, keepRich = TRUE, filename = NULL,
  cutoff = 0.05)
}
\arguments{
\item{df}{DGE files (DESeq2 result files) or vector contains gene names}

\item{RO_FILE}{GO annotation data}

\item{gene.cutoff}{the cut-off value for select DEGs (default: 0.01)}

\item{padj.method}{p value adjust method (default: BH)}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{filename}{output filename}

\item{cutoff}{cutoff value for filtering significant terms (default: 0.05)}
}
\description{
Reactome Pathway Enrichment analysis function for plant
}
\author{
Kai Guo
}
