% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RE.R
\name{RE}
\alias{RE}
\title{Reactome Pathway Enrichment analysis function}
\usage{
RE(
  df,
  RO_FILE,
  keytype = "SYMBOL",
  species = "human",
  padj.method = "BH",
  minSize = 2,
  maxSize = 500,
  keepRich = TRUE,
  filename = NULL,
  cutoff = 0.05
)
}
\arguments{
\item{df}{DGE files (DESeq2 result files) or vector contains gene names}

\item{RO_FILE}{Reactome annotation data}

\item{species}{species name}

\item{padj.method}{p value adjust method (default: BH)}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{filename}{output filename}

\item{cutoff}{cutoff value for filtering significant terms (default: 0.05)}

\item{gene.cutoff}{the cut-off value for select DEGs (default: 0.01)}
}
\description{
Reactome Pathway Enrichment analysis function
}
\author{
Kai Guo
}
