% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setEnrichrSite}
\alias{setEnrichrSite}
\title{Set Enrichr Website}
\usage{
setEnrichrSite(site)
}
\arguments{
\item{site}{site requested}
}
\value{
Changes Enrichr Website connection
}
\description{
Set Enrichr Website
}
\details{
Set Enrichr Website
}
\author{
Alexander Blume
}
