% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrodata.R
\name{showAvailableplantsRO}
\alias{showAvailableplantsRO}
\title{show avaliable plant Reactome annotation data function}
\usage{
showAvailableplantsRO()
}
\description{
show avaliable plant Reactome annotation data function
}
\author{
Kai Guo
}
