% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrodata.R
\name{showAvailableRO}
\alias{showAvailableRO}
\title{show avaliable Reactome annotation data except plant}
\usage{
showAvailableRO()
}
\description{
show avaliable Reactome annotation data except plant
}
\author{
Kai Guo
}
