% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{showensemble}
\alias{showensemble}
\title{Show avaliable annotation except plant based on ENSEMBLE}
\usage{
showensemble(host = "uswest.ensembl.org")
}
\arguments{
\item{host}{the ensemble API host,for plant you can use plants.ensembl.org and for human and other species you can use uswest.ensembl.org}

\item{species}{the sepcies you want to search}
}
\description{
Show avaliable annotation except plant based on ENSEMBLE
}
\author{
Kai Guo
}
