% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{showplant}
\alias{showplant}
\title{Show avaliable plant annotation based on ENSEMBLE}
\usage{
showplant(host = "plants.ensembl.org")
}
\arguments{
\item{host}{the ensemble API host,for plant you can use plants.ensembl.org and for human and other species you can use uswest.ensembl.org}

\item{species}{the species you want to search}

\item{ann_type}{the type of function annotation you want get from ensemble}
}
\description{
Show avaliable plant annotation based on ENSEMBLE
}
\author{
Kai Guo
}
