% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF.R
\name{TFenrich}
\alias{TFenrich}
\title{TFenrich: in fact it can enrich everything,just prepare the right database.}
\usage{
TFenrich(genelist, genetype, TFfile = "TF.info")
}
\arguments{
\item{genelist:a}{gene list vector.}

\item{genetype:a}{string.}

\item{TFfile:}{just like \code{GO.info}.}
}
\value{
return a S3 enrich result
}
\description{
TFenrich: in fact it can enrich everything,just prepare the right database.
}
\examples{
genelist <- read.delim("ST1.txt")
TFenrich(genelist,"TF1","data/TF.info")
}
