% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.onAttach}
\alias{.onAttach}
\title{onLoad hook to setup package options}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{(Required). Library name}

\item{pkgname}{(Required). Package name}
}
\description{
onLoad hook to setup package options
}
\details{
onLoad hook to setup package options and to check connection to website
}
\author{
Wajid Jawaid \email{wj241@alumni.cam.ac.uk}
}
