% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichbar}
\alias{enrichbar}
\title{Display enrichment result By using barchart}
\usage{
enrichbar(
  resultFis,
  top = 50,
  pvalue.cutoff = 0.05,
  padj.cutoff = NULL,
  low = "lightpink",
  high = "red",
  order = FALSE,
  horiz = FALSE,
  fontsize.x = 10,
  fontsize.y = 10,
  fontsize.text = 3,
  angle = 75,
  usePadj = TRUE,
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{resultFis}{Ennrichment analysis result data.frame}

\item{top}{Number of Terms you want to display}

\item{pvalue.cutoff}{the cut-off value for selecting Term}

\item{padj.cutoff}{the padj cut-off value for selecting Term}

\item{low}{color used for small value}

\item{high}{color used for large value}

\item{horiz}{show horiz or not (default: FALSE)}

\item{fontsize.x}{fontsize for x axis}

\item{fontsize.y}{fontsize for y axis}

\item{filename}{output filename}

\item{width}{width for output file}

\item{height}{height for output file}
}
\description{
Display enrichment result By using barchart
}
\author{
Kai Guo
}
