% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichdot}
\alias{enrichdot}
\title{Display enrichment result By using dotchart}
\usage{
enrichdot(resultFis, top = 50, pvalue.cutoff = 0.05, order = FALSE,
  padj.cutoff = NULL, fontsize.x = 10, fontsize.y = 10, usePadj = TRUE,
  filename = NULL)
}
\arguments{
\item{resultFis:}{Ennrichment analysis result data.frame}

\item{top:}{Number of Terms you want to display}

\item{pvalue.cutoff:}{the cut-off value for selecting Term}

\item{padj.cutoff:}{the padj cut-off value for selecting Term}
}
\description{
Display enrichment result By using dotchart
}
\author{
Kai Guo
}
