% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrapper.R
\name{enrichGeneList}
\alias{enrichGeneList}
\title{Perform functional enrichment on a set of genes.}
\usage{
enrichGeneList(gene.list, databases = db.list)
}
\arguments{
\item{gene.list}{a list of gene symbols}

\item{databases}{a list of Enrichr-fronted databases, as mentioned above}
}
\description{
This function interacts with Enrichr's REST API in order to perform functional enrichment of a single
set of genes, for a set of specified databases which are already fronted by Enrichr.
Databases are specified as seen in the web interface, with underscores for spaces
(e.g. "WikiPathways_2016", "KEGG_2016", "GO_Biological_Process"). There's no way to query Enrichr
to get these database names, so they can't be provided as options. You'll just have to guess. Sorry :/
}
\keyword{Enrichr}
\keyword{enrichment}
\keyword{functional}

