% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enrichr}
\alias{enrichr}
\title{Gene enrichment using Enrichr}
\usage{
enrichr(genes, databases = NULL, background = NULL)
}
\arguments{
\item{genes}{(Required). Character vector of Entrez gene symbols as input. A data.frame
of gene symbols in first column is also acceptable, optionally a score denoting the
degree of membership between 0 and 1 in the second column.}

\item{databases}{(Required). Character vector of databases to search.
See https://maayanlab.cloud/Enrichr/ for available databases.}

\item{background}{(Optional). Character vector of Entrez gene symbols to be used as
background. A data.frame of gene symbols in first column is also acceptable.
Enrichment analysis with background genes is only available on the main site (Enrichr).}
}
\value{
Returns a list of data.frame of enrichment terms, p-values, ...
}
\description{
Gene enrichment using Enrichr
}
\details{
Gene enrichment using Enrichr
}
\examples{
dbs <- listEnrichrDbs()
dbs <- c("GO_Molecular_Function_2018", "GO_Cellular_Component_2018",
         "GO_Biological_Process_2018")
enriched <- enrichr(c("Runx1", "Gfi1", "Gfi1b", "Spi1", "Gata1", "Kdr"), dbs)
}
\author{
Wajid Jawaid \email{wj241@alumni.cam.ac.uk}
}
