% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE.R
\name{GE.plot}
\alias{GE.plot}
\title{Display GO enrichment result}
\usage{
GE.plot(
  resultFis,
  top = 50,
  pvalue.cutoff = 0.05,
  order = FALSE,
  horiz = FALSE,
  low = "lightpink",
  high = "red",
  font.x = "bold",
  font.y = "bold",
  fontsize.x = 10,
  fontsize.y = 10,
  fontsize.text = 3,
  angle = 75,
  padj.cutoff = NULL,
  usePadj = TRUE,
  filename = NULL,
  width = 10,
  height = 8
)
}
\arguments{
\item{order}{order bar or not}

\item{horiz}{horizontal plot or not (default: FALSE)}

\item{low}{color used for small value}

\item{high}{color used for large value}

\item{fontsize.x}{fontsize for x axis}

\item{fontsize.y}{fontsize for y axis}

\item{angle}{angle for x ticks label}

\item{usePadj}{use adjust pvalue or not}

\item{filename}{output filename}

\item{width}{width for output file}

\item{height}{height for output file}

\item{resultFis:}{GO ennrichment analysis result data.frame}

\item{top:}{Number of Terms you want to display}

\item{filename:}{output filename}

\item{pvalue.cutoff:}{the cut-off value for selecting Term}

\item{padj.cutoff:}{the padj cut-off value for selecting Term}
}
\description{
Display GO enrichment result
}
\author{
Kai Guo
}
