% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE.R
\name{GE}
\alias{GE}
\title{GO Enrichment analysis function}
\usage{
GE(df, GO_FILE, OP = "BP", gene.cutoff = 0.01, minSize = 2,
  maxSize = 500, keepRich = TRUE, filename = NULL,
  padj.method = "BH", cutoff = 0.05)
}
\arguments{
\item{df:}{DGE files (DESeq2 result files) or vector contains gene names}

\item{GO_FILE:}{GO annotation data}

\item{filenam:}{output filename}

\item{gene.cutoff:}{the cut-off value for select DGE}
}
\description{
GO Enrichment analysis function
}
\author{
Kai Guo
}
