% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEnrichr}
\alias{getEnrichr}
\title{Helper function for HTTP methods GET and POST}
\usage{
getEnrichr(method = "GET", url, ...)
}
\arguments{
\item{method}{(Required). HTTP method. Default is \code{"GET"}}

\item{url}{(Required). URL address requested}

\item{...}{(Optional). Additional parameters to pass to GET/POST}
}
\value{
same as GET/POST
}
\description{
Helper function
}
\details{
Helper function for HTTP methods GET and POST
}
\author{
Wajid Jawaid \email{wj241@alumni.cam.ac.uk}

I-Hsuan Lin \email{i-hsuan.lin@manchester.ac.uk}
}
