% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEnrichr}
\alias{getEnrichr}
\title{Helper function for GET}
\usage{
getEnrichr(url, ...)
}
\arguments{
\item{url}{url address requested}

\item{...}{Additional parameters to pass to GET}
}
\value{
same as GET
}
\description{
Helper function
}
\details{
Helper function for GET
}
\author{
Wajid Jawaid
}
