% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEnrichr}
\alias{getEnrichr}
\title{Helper function for GET}
\usage{
getEnrichr(url, ...)
}
\arguments{
\item{url}{(Required). URL address requested}

\item{...}{(Optional). Additional parameters to pass to GET}
}
\value{
same as GET
}
\description{
Helper function
}
\details{
Helper function for GET
}
\author{
Wajid Jawaid \email{wj241@alumni.cam.ac.uk}
}
