% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE.R
\name{IE.ensemble}
\alias{IE.ensemble}
\title{InterPro Enrichment analysis function}
\usage{
IE.ensemble(df, IO_FILE, gene.cutoff = 0.01, minSize = 2,
  maxSize = 500, keepRich = TRUE, padj.method = "BH",
  filename = NULL, cutoff = 0.05)
}
\arguments{
\item{df:}{DGE files (DESeq2 result files) or vector contains gene names}

\item{IO_FILE:}{InterPro annotation data, you can get the data by using makeesanno function}

\item{filename:}{output filename}

\item{gene.cutoff:}{DGE singificant cutoff value}
}
\description{
InterPro Enrichment analysis function
}
\author{
Kai Guo
}
