% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE.R
\name{IE.plant}
\alias{IE.plant}
\title{InterPro Enrichment analysis function for plant}
\usage{
IE.plant(df, IO_FILE, gene.cutoff = 0.01, minSize = 2, maxSize = 500,
  keepRich = TRUE, padj.method = "BH", filename = NULL,
  cutoff = 0.05)
}
\arguments{
\item{df:}{DGE files (DESeq2 result files) or vector contains gene names}

\item{IO_FILE:}{InterPro annotation data, you can get the data by using makeplantann function}

\item{filename:}{output filename}

\item{gene.cutoff:}{DGE singificant cutoff value}
}
\description{
InterPro Enrichment analysis function for plant
}
\author{
Kai Guo
}
