% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KE.R
\name{KE}
\alias{KE}
\title{KEGG Pathway Enrichment analysis function}
\usage{
KE(df, KO_FILE, filename = NULL, gene.cutoff = 0.01, minSize = 2,
  maxSize = 500, keepRich = TRUE, padj.method = "BH",
  cutoff = 0.05)
}
\arguments{
\item{df:}{DGE files (DESeq2 result files) or vector contains gene names}

\item{KG_FILE:}{KEGG Pathway annotation data}

\item{filenam:}{output filename}

\item{cutoff:}{DGE singificant cutoff value}
}
\description{
KEGG Pathway Enrichment analysis function
}
