% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmap_m.R
\name{lheatmap}
\alias{lheatmap}
\title{Plot compare heatmap of Enrichment result among DEG groups}
\usage{
lheatmap(
  rhs,
  top = 50,
  colnames = NULL,
  xsize = 6,
  ysize = 6,
  padj = NULL,
  horizontal = FALSE,
  pval = 0.05,
  returnData = FALSE,
  ...
)
}
\arguments{
\item{top}{the number of Terms you want to display}

\item{colnames}{the compare DEG group names}

\item{xsize}{cex of group name}

\item{ysize}{cex of Terms name}

\item{rhslist}{of enrchment analysis result among DEG groups}
}
\description{
Plot compare heatmap of Enrichment result among DEG groups
}
\author{
Kai Guo
}
