% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org.R
\name{makeGOdat}
\alias{makeGOdat}
\title{make GO annotation data function}
\usage{
makeGOdat(species = "human", keytype = "ENTREZID")
}
\arguments{
\item{species:}{you can check the support species by using showData()}

\item{keytype:}{the gene type}
}
\description{
make GO annotation data function
}
\author{
Kai Guo
}
