% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmap_m.R
\name{mnetmap}
\alias{mnetmap}
\title{plot combine network of Terms}
\usage{
mnetmap(df, gores = NULL, kores = NULL, rores = NULL, pfres = NULL,
  itres = NULL, top = NULL, top.display = NULL,
  pvalue.cutoff = 0.05, padj.cutoff = NULL, visNet = F,
  weightcut = 0.2, useTerm = TRUE, layout = NULL,
  vertex.label.cex = 0.5, gnet = FALSE, vertex.node.shape = NULL,
  ...)
}
\arguments{
\item{vertex.node.shape}{vector of shape and names of the vector should be the terms (default: 20)}

\item{df:}{DGE files (DESeq2 result files) or vector contains gene names}

\item{gores:}{GO Enrichment analsyis result}

\item{kores:}{KEGG Enrichment analsyis result}

\item{rores:}{Reactome Pathway Enrichment analsyis result}

\item{pfres:}{PFAM Enrichment analysis result result}

\item{itres:}{InterPro Enrichment analysis result}

\item{pvalue.cutoff:}{the cut-off P value for selecting significant Terms}

\item{padj.cutoff:}{the cut-off P adjust value for selecting significant Terms}

\item{weightcut:}{the weight cut value for remove edges}

\item{useTerm:}{use the Term description or not(defalut:TRUE)}

\item{writeCyt:}{export file for Cyt software}

\item{vertex.label.color:}{color of label(default:black)}

\item{vertex.label.cex:}{size of label(default:0.5)}

\item{layout:}{layout format (defult:layout.fruchterman.reingold)}

\item{visNet:}{use VisNetwork method to display network(default:FALSE)}

\item{top:}{number of Terms you want to display(default:the total number of all Significant number)}
}
\description{
plot combine network of Terms
}
\author{
Kai Guo
}
