% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{printEnrich}
\alias{printEnrich}
\title{Print Enrichr output to text file.}
\usage{
printEnrich(data, file, sep = "\\t", columns = c(2, 3, 6))
}
\arguments{
\item{data}{Output from Enrichr function.}

\item{file}{Name of output file.}

\item{sep}{Default TAB. How to separate fields.}

\item{columns}{Columns from each entry of data.
1-"Index", 2-"Name", 3-"Adjusted_P-value", 4-"Z-score"
5-"Combined_Score", 6-"Genes", 7-"Overlap_P-value"}
}
\value{
Produces file.
}
\description{
Print Enrichr output.
}
\details{
Print Enrichr output to text file.
}
\author{
Wajid Jawaid
}
