% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RE.R
\name{RE}
\alias{RE}
\title{Reactome Pathway Enrichment analysis function}
\usage{
RE(df, RO_FILE, keytype = "SYMBOL", species = "human", padj.method = "BH",
  minSize = 2, maxSize = 500, keepRich = TRUE, filename = NULL)
}
\arguments{
\item{df:}{DGE files (DESeq2 result files) or vector contains gene names}

\item{RO_FILE:}{Reactome Pathway annotation data, you can get the data by use makeROdata or makePlantROdat function}

\item{filename:}{output filename}

\item{gene.cutoff:}{DGE singificant cutoff value}
}
\description{
Reactome Pathway Enrichment analysis function
}
\author{
Kai Guo
}
