% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richplot.R
\name{richplot}
\alias{richplot}
\title{richplot}
\usage{
richplot(resultFis, top = 50, pvalue.cutoff = 0.05,
  padj.cutoff = NULL, usePadj = TRUE, useTerm = TRUE,
  writeCyt = FALSE, cytoscapeFile = "network-file-for-cytoscape.txt",
  label.color = "black", label.size = 2, node.shape = NULL,
  layout = layout.fruchterman.reingold, savefig = FALSE,
  filename = "network", width = 7, height = 7, node.alpha = 0.7)
}
\arguments{
\item{resultFis}{Enrichment results}

\item{top}{number of terms to show (default: 50)}

\item{pvalue.cutoff}{cutoff p value for enrichment result}

\item{padj.cutoff}{cutoff p adjust value for enrichment result}

\item{useTerm}{use terms for nodes (default: TRUE)}

\item{writeCyt}{write out the cytoscape file}

\item{cytoscapeFile}{output cytoscape File}

\item{label.color}{label color}

\item{label.size}{label size}

\item{node.shape}{vector of shape and names of the vector should be the terms (default: 20)}

\item{layout}{layout method}

\item{savefig}{save figures or not}

\item{filename}{output figure name}

\item{width}{width for output figure}

\item{height}{height for output figure}

\item{node.alpha}{alpha-transparency scales}
}
\description{
plot the sigificant terms and shared genes with network format
}
